requirements.txt
attrs==19.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dash==0.39.0
dash-core-components==0.44.0
dash-html-components==0.14.0
dash-renderer==0.20.0
dash-table==3.6.0
decorator==4.3.2
Flask==1.0.2
Flask-Compress==1.4.0
idna==2.8
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.1
jupyter-core==4.4.0
lxml==4.3.2
MarkupSafe==1.1.1
nbformat==4.4.0
numpy==1.16.2
pandas==0.24.1
pandas-datareader==0.7.0
plotly==3.7.0
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
retrying==1.3.3
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1
